Re-organzie code - extract file related code
